requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
aiohttp==3.9.1
protobuf==4.25.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-socketio==5.11.0
aioredis==2.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
httpx==0.26.0
loguru==0.7.2
python-dotenv==1.0.0
brotli==1.1.0
